requirements.txt
matplotlib~=3.5.1
numpy~=1.22.0
pyserial~=3.5
pandas~=1.3.5